requirements.txt
flask
gunicorn
Pillow
python-pptx
requests
serpapi
flask-cors
groq